fix: cast scores to float32 before quantile() call
quantile() requires float32/float64; Mistral attention weights are float16.

+++ src/adaptive/cache_manager.py
@@ -114,7 +114,7 @@ class AdaptiveCacheManager:
 
         if n_compressible > 0:
             compressible_scores = scores[compress_start:compress_end]
-            threshold = compressible_scores.quantile(self.compress_ratio)
+            threshold = compressible_scores.float().quantile(self.compress_ratio)
 
             for i in range(compress_start, compress_end):
                 if i not in self._compressed and compressible_scores[i - compress_start] <= threshold:
